Trim_network function and visualization rework
Trim_network allows for selecting only a subset of nodes. The option "expand" includes also the nodes that are parents or children of the given subset of interest nodes.

The visualization has been reworked and no longer depends on dbnR. In addition, a function to visualize spatio_temporal networks was added